refactoring of the select and deselect services
